pyproject.toml
[project]
name = "finance-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "dotenv>=0.9.9",
    "plotly>=6.3.0",
    "praw>=7.8.1",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.5",
    "streamlit>=1.49.1",
]
